+++ app/Console/Kernel.php
@@ -39,7 +39,9 @@ class Kernel extends ConsoleKernel
 
         if(config('import.instagram.enabled')) {
             $schedule->command('app:transform-imports')->everyFourMinutes();
-            $schedule->command('app:import-upload-garbage-collection')->everyFiveMinutes();
+            $schedule->command('app:import-upload-garbage-collection')->hourlyAt(51);
+            $schedule->command('app:import-remove-deleted-accounts')->hourlyAt(37);
+            $schedule->command('app:import-upload-clean-storage')->twiceDailyAt(1, 13, 32);
         }
     }